+++ tests/test_plugin.py
@@ -1,0 +1,8 @@
+import unittest
+
+from octofarm_companion import __plugin_load__
+
+
+class TestPlugin(unittest.TestCase):
+    def test_load(self):
+        __plugin_load__()